WIP WIP, tried react-bootstrap-table package
repeat of ongoing issue with react and third-party components that work in their demos but not in this project code. the problem could be due in part to not knowing enough about react/jsx/etc to extract exactly what's needed from demos. there are also issues that seem to be linked to how well (or not) specific versions of the dependancies play with each other.